refactor: extract types from data module
