+++ expertsystem/amplitude/canonical_decay.py
@@ -2,8 +2,9 @@
 
 from typing import Any, Callable, List, Optional
 
-from expertsystem.data import InteractionProperties, ParticleWithSpin, Spin
+from expertsystem.data import Spin
 from expertsystem.topology import StateTransitionGraph
+from expertsystem.types import InteractionProperties, ParticleWithSpin
 
 from .helicity_decay import (
     HelicityAmplitudeGenerator,
