+++ expertsystem/amplitude/helicity_decay.py
@@ -2,11 +2,12 @@
 import logging
 from typing import Dict, List, Optional, Tuple, Union
 
-from expertsystem.data import ParticleCollection, ParticleWithSpin, Spin
+from expertsystem.data import ParticleCollection, Spin
 from expertsystem.state.properties import (
     perform_external_edge_identical_particle_combinatorics,
 )
 from expertsystem.topology import StateTransitionGraph, Topology
+from expertsystem.types import ParticleWithSpin
 
 from .model import (
     AmplitudeModel,
