+++ expertsystem/solving/conservation_rules.py
@@ -20,7 +20,8 @@ from typing import Any, Callable, List, Optional, Set, Tuple, Type, Union
 import attr
 
 from expertsystem._utils import arange
-from expertsystem.data import EdgeQuantumNumbers, NodeQuantumNumbers, Spin
+from expertsystem.data import Spin
+from expertsystem.types import EdgeQuantumNumbers, NodeQuantumNumbers
 
 
 def is_boson(spin_magnitude: float) -> bool:
