+++ tests/unit/solving/conservation_rules/test_c_parity.py
@@ -2,12 +2,13 @@ from itertools import product
 
 import pytest
 
-from expertsystem.data import EdgeQuantumNumbers, NodeQuantumNumbers, Parity
+from expertsystem.data import Parity
 from expertsystem.solving.conservation_rules import (
     CParityConservation,
     CParityEdgeInput,
     CParityNodeInput,
 )
+from expertsystem.types import EdgeQuantumNumbers, NodeQuantumNumbers
 
 # Currently need to cast to the proper Edge/NodeQuantumNumber type, see
 # https://github.com/ComPWA/expertsystem/issues/255
